chore: migrate Next.js config to MJS with redirects
