+++ next.config.mjs
@@ -4,6 +4,20 @@ import { withContentlayer } from 'next-contentlayer'
 const nextConfig = {
   reactStrictMode: true,
   experimental: { typedRoutes: false },
+  async redirects() {
+    return [
+      {
+        source: '/kontakt',
+        destination: '/contact',
+        statusCode: 301,
+      },
+      {
+        source: '/dla-ksiegowych',
+        destination: '/ksiegowi',
+        statusCode: 301,
+      },
+    ]
+  },
 }
 
 export default withContentlayer(nextConfig)
